v1.13f - reworked arg processing & other fixes

+++ detect_rescan.sh
@@ -30,14 +30,16 @@ output() {
     echo "detect_rescan: $*"
 }
  
-output "Starting Detect Rescan wrapper v1.12c"
+output "Starting Detect Rescan wrapper v1.13f"
 
 DETECT_TMP=$(mktemp -u)
 TEMPFILE=$(mktemp -u)
 TEMPFILE2=$(mktemp -u)
 LOGFILE=$(mktemp -u)
 
-ACTION_ARGS=( "--blackduck.timeout" "--detect.force.success" "--detect.notices.report" "--detect.policy.check.fail.on.severities" "--detect.risk.report.pdf" "--detect.wait.for.results" )
+# ACTION_ARGS=--blackduck.timeout=*|--detect.force.success=*|--detect.notices.report=*|--detect.policy.check.fail.on.severities=*|--detect.risk.report.pdf=*|--detect.wait.for.results=*
+# UNSUPPORTED_ARGS='--detect.blackduck.signature.scanner.snippet.matching=*|--detect.blackduck.signature.scanner.upload.source.mode=*|--detect.blackduck.signature.scanner.copyright.search=*|--detect.blackduck.signature.scanner.license.search=*|--detect.binary.scan.*'
+
 API_TOKEN=$BLACKDUCK_API_TOKEN
 BD_URL=$BLACKDUCK_URL
 YML=
@@ -63,6 +65,8 @@ MODE_RESET=0
 JQTEMPDIR=
 JQ=
 
+UNSUPPORTED=0
+
 BOM_FILES=()
 BOM_HASHES=()
 PREV_FILES=()
@@ -73,6 +77,12 @@ error() {
     echo "detect_rescan: ERROR: $*" >$LOGFILE
     cat $LOGFILE
     end 1
+}
+
+error2() {
+    echo "detect_rescan: ERROR: $*" >$LOGFILE
+    cat $LOGFILE
+    end 2
 }
 
 end() {
@@ -96,6 +106,24 @@ msg() {
     if [ $MODE_QUIET -eq 0 ]
     then
         output "$*"
+    fi
+}
+
+encode_url() {
+    if [ ! -z "$1" ]
+    then
+        echo ${*} | sed -e 's:/:%2F:g' -e 's/ /%20/g' -e 's/\[/%5B/g' -e 's/\]/%5D/g' -e 's/{/%7B/g' -e 's/}/%7D/g' -e 's/(/%28/g' -e 's/)/%29/g' -e 's/"//g'
+    else
+        cat - | sed -e 's:/:%2F:g' -e 's/ /%20/g' -e 's/\[/%5B/g' -e 's/\]/%5D/g' -e 's/{/%7B/g' -e 's/}/%7D/g' -e 's/(/%28/g' -e 's/)/%29/g' -e 's/\"//g'
+    fi
+}
+
+escape_string() {
+    if [ ! -z "$1" ]
+    then
+        echo ${*} | sed -e 's/ /\\ /g' -e 's/"//g'
+    else
+        cat - | sed -e 's/ /\\ /g' -e 's/"//g'
     fi
 }
 
@@ -152,133 +180,7 @@ prereqs() {
     return $ret
 }
 
-process_args() {
-    local UNSUPPORTED=0
-    local prevarg=
-    DETARGS=
-    for arg in $*
-    do
-        if [[ $arg == --blackduck.offline.mode=* ]]
-        then
-            continue
-        elif [[ $arg == --detect.blackduck.signature.scanner.host.url=* ]]
-        then
-            continue
-        elif [[ $arg == --blackduck.api.token=* ]]
-        then
-            debug "process_args(): BLACKDUCK_API_TOKEN identified from command line option"
-            API_TOKEN=$(echo $arg | cut -f2 -d=)==
-        elif [[ $arg == --blackduck.url=* ]]
-        then
-            debug "process_args(): BLACKDUCK_URL identified from command line option"
-            BD_URL=$(echo $arg | cut -f2 -d=)
-        elif [[ $arg == --detect.project.name=* ]]
-        then
-            DETECT_PROJECT=1
-        elif [[ $arg == --detect.blackduck.signature.scanner.snippet.matching=* ]]
-        then
-            debug "process_args(): --detect.blackduck.signature.scanner.snippet.matching option identified"
-            UNSUPPORTED=1
-        elif [[ $arg == --detect.blackduck.signature.scanner.upload.source.mode=* ]]
-        then
-            debug "process_args(): --detect.blackduck.signature.scanner.upload.source.mode option identified"
-            UNSUPPORTED=1
-        elif [[ $arg == --detect.blackduck.signature.scanner.copyright.search=* ]]
-        then
-            debug "process_args(): --detect.blackduck.signature.scanner.copyright.search option identified"
-            UNSUPPORTED=1
-        elif [[ $arg == --detect.blackduck.signature.scanner.license.search=* ]]
-        then
-            debug "process_args(): --detect.blackduck.signature.scanner.license.search option identified"
-            UNSUPPORTED=1
-        elif [[ $arg == --detect.binary.scan.* ]]
-        then
-            debug "process_args(): --detect.binary.scan.* identified"
-            UNSUPPORTED=1
-        elif [[ $arg == --detect.project.version.name=* ]]
-        then
-            DETECT_VERSION=1
-        elif [[ $arg == --spring.profiles.active=* ]]
-        then
-            local YML="application-$(echo $arg | cut -f2 -d=).yml"
-            if [ ! -r $YML ]
-            then
-                YML=
-            fi
-        fi
-        for opt in ${ACTION_ARGS[@]}
-        do
-            if [[ $arg == ${opt}* ]]
-            then
-                DETECT_ACTION=1
-                msg "Detect Action identified - will rerun Detect after upload and scan completion"
-                debug "process_args(): Identified action argument $arg"
-            fi
-        done
-
-        if [ "$arg" == "--report" ]
-        then
-            debug "process_args(): MODE_REPORT set"
-            MODE_REPORT=1
-        elif [ "$arg" == "--quiet" ]
-        then
-            debug "process_args(): MODE_QUIET set"
-            MODE_QUIET=1
-        elif [ "$arg" == "--markdown" ]
-        then
-            debug "process_args(): MODE_MARKDOWN set"
-            MODE_MARKDOWN=1
-        elif [ "$arg" == "--file" ]
-        then
-            debug "process_args(): MODE_PREVFILE set"
-            MODE_PREVFILE=1
-        elif [ "$arg" == "--testxml" ]
-        then
-            debug "process_args(): MODE_TESTXML set"
-            MODE_TESTXML=1
-        elif [ "$arg" == "--reset" ]
-        then
-            debug "process_args(): MODE_RESET set"
-            MODE_RESET=1
-        elif [[ $arg == --detectscript=* ]]
-        then
-            DETECT_SCRIPT=$(echo $arg | cut -f2 -d=)
-            debug "process_args(): DETECT_SCRIPT set to $DETECT_SCRIPT"
-            if [ ! -r "$DETECT_SCRIPT" ]
-            then
-                error "Detect script $DETECT_SCRIPT does not exist"
-            fi
-        elif [[ $arg == --sigtime=* ]]
-        then
-            SIGTIME=$(echo $arg | cut -f2 -d=)
-            debug "process_args(): SIGTIME set to $SIGTIME"
-        elif [[ $arg == --curlopts=* ]]
-        then
-            CURLOPTS=$(echo $arg | cut -f2 -d=)
-        elif [[ $arg == --* ]]
-        then
-            if [ ! -z "$prevarg" ]
-            then
-                DETARGS="$DETARGS '$prevarg'"
-            fi
-            prevarg=$arg
-        else
-            prevarg="$prevarg $arg"
-        fi
-        if [[ $prevarg == --detect.source.path=* ]]
-        then
-            SCANLOC=$(echo $prevarg | cut -f2 -d=)
-            SCANLOC=$(cd "$SCANLOC" 2>/dev/null; pwd)
-        fi
-    done
-    if [[ $prevarg == --detect.source.path=* ]]
-    then
-        SCANLOC=$(echo $prevarg | cut -f2 -d=)
-        SCANLOC=$(cd "$SCANLOC"; pwd)
-    fi
-    debug "process_args(): SCANLOC set to $SCANLOC"
-
-    DETARGS="$DETARGS '$prevarg'"
+check_env() {
     #
     # Check Environment variables
     if [ "$DETECT_BLACKDUCK_SIGNATURE_SCANNER_SNIPPET_MATCHING" == "true" ]
@@ -466,6 +368,7 @@ proc_bom_files() {
         fi
         CKSUM=$(cat $bom | grep -v 'spdx:created' | grep -v 'uuid:' | sort | cksum | cut -f1 -d' ')
         FILE=$(basename $bom)
+        debug "proc_bom_files(): Checksum for '$FILE' is '$CKSUM'"
         BOM_FILES+=("${FILE}")
         BOM_HASHES+=("${CKSUM}")
         ((COUNT++))
@@ -623,23 +526,29 @@ get_project() {
     #Get  projects $1=projectname
     debug "get_project(): ARG1=$1"
     #local SEARCHPROJ=$(echo ${1} | sed -e 's:/:%2F:g' -e 's/ /+/g')
-    local SEARCHPROJ=$(echo ${1} | sed -e 's:/:%2F:g' -e 's/ /%20/g' -e 's/\[/%5B/g' -e 's/\]/%5D/g' )
+    local SEARCHPROJ=$(encode_url ${1})
     local MYURL="${BD_URL}/api/projects?q=name:${SEARCHPROJ}"
     debug "get_project(): API_URL=$MYURL"
 
     api_call "$MYURL" 'application/vnd.blackducksoftware.project-detail-4+json'
     if [ $? -ne 0 ]
     then
+        debug "get_project(): API error - returning early"
         return 1
     fi
 
-    local PROJNAMES=$($JQ -r '[.items[].name]|@csv' $TEMPFILE 2>/dev/null| sed -e 's/ /%20/g' -e 's/\"//g' -e 's:/:%2F:g'  -e 's/\[/%5B/g' -e 's/\]/%5D/g')
-    local PROJURLS=$($JQ -r '[.items[]._meta.href]|@csv' $TEMPFILE 2>/dev/null| sed -e 's/\"//g')
+    local PROJNAMES=$($JQ -r '[.items[].name]|@csv' $TEMPFILE 2>/dev/null| encode_url )
+    local PROJURLS=$($JQ -r '[.items[]._meta.href]|@csv' $TEMPFILE 2>/dev/null| sed -e 's/\"//g' )
+    debug "get_project(): PROJNAMES=$PROJNAMES"
+    debug "get_project(): PROJURLS=$PROJURLS"
+
     local PROJNUM=1
     local FOUNDNUM=0
     local IFS=,
     for PROJ in $PROJNAMES
     do
+        debug "get_project(): PROJ='$PROJ' SEARCHPROJ='$SEARCHPROJ'"
+
         if [ "$PROJ" == "$SEARCHPROJ" ]
         then
             FOUNDNUM=$PROJNUM
@@ -649,11 +558,13 @@ get_project() {
     done
     IFS=
 
+    debug "get_project(): Found $FOUNDNUM projects"
     if [ $FOUNDNUM -eq 0 ]
     then
         return 0
     fi
 
+    debug "get_project(): PROJURLS is '$PROJURLS'"
     RETURL=$(echo $PROJURLS | cut -f $FOUNDNUM -d ,)
     debug "get_project(): returning project URL $RETURL"
     echo $RETURL
@@ -662,18 +573,22 @@ get_project() {
 
 get_version() {
     # Get Version  - $1 = PROJURL
-    local VERNAME=$(echo $2 | sed -e 's:/:%2F:g' -e 's/ /%20/g' -e 's/\[/%5B/g' -e 's/\]/%5D/g')
+    local VERNAME=$(encode_url ${2} )
     local API_URL="${1//\"}/versions?versionName%3A${VERNAME}"
+    debug "get_version(): version URL is '$API_URL'"
     #local SEARCHVERSION="${2// /_}"
     #echo "get_version: SEARCHVERSION=$SEARCHVERSION" >&2
     api_call "${API_URL}" 'application/vnd.blackducksoftware.project-detail-4+json'
     if [ $? -ne 0 ]
     then
+        debug "get_version(): API error"
         return 1
     fi
 
-    local VERNAMES=$($JQ -r '[.items[].versionName]|@csv' $TEMPFILE 2>/dev/null | sed -e 's/ /%20/g' -e 's/\"//g' -e 's:/:%2F:g'  -e 's/\[/%5B/g' -e 's/\]/%5D/g')
-    local VERURLS=$($JQ -r '[.items[]._meta.href]|@csv' $TEMPFILE 2>/dev/null | sed -e 's/\"//g')
+    local VERNAMES=$($JQ -r '[.items[].versionName]|@csv' $TEMPFILE 2>/dev/null | encode_url )
+    local VERURLS=$($JQ -r '[.items[]._meta.href]|@csv' $TEMPFILE 2>/dev/null | sed -e 's/\"//g' )
+    debug "get_version(): VERNAMES=$VERNAMES"
+    debug "get_version(): VERURLS=$VERURLS"
     local VERNUM=1
     local FOUNDVERNUM=0
     local IFS=,
@@ -691,6 +606,7 @@ get_version() {
     
     if [ $FOUNDVERNUM -eq 0 ]
     then
+        debug "get_version(): 0 versions found from project"
         return 0
     fi
 
@@ -759,18 +675,19 @@ wait_for_bom_completion() {
                 STATUS='true'
             fi
         fi
-        
+
         if [ "$STATUS" == "true" ]
         then
             debug "wait_for_bom_completion(): upToDate status returned true"
-            break
+            echo
+            return 0
         fi
         echo -n '.'
         sleep 15
         ((loop++))
     done
     echo
-    return 0
+    return 1
 }
 
 wait_for_scans() {
@@ -814,7 +731,7 @@ wait_for_scans() {
         echo -n '.'
         sleep 15
     done
-    return 0
+    return 1
 }
 
 check_sigscan() {
@@ -912,22 +829,25 @@ proc_sigscan() {
 }
 
 cleanup() {
-    if [ ! -z "$RUNDIR" ]
+    if [ -z "$DEBUG" ]
     then
-        if [ -d "$RUNDIR/bdio" ]
+        if [ ! -z "$RUNDIR" ]
         then
-            rm -rf "$RUNDIR/bdio"
-            msg "Deleting $RUNDIR/bdio"
-        fi
-        if [ -d "$RUNDIR/extractions" ]
-        then
-            rm -rf "$RUNDIR/extractions"
-            msg "Deleting $RUNDIR/extractions"
-        fi
-        if [ -d "$RUNDIR/scan" ]
-        then
-            rm -rf "$RUNDIR/scan"
-            msg "Deleting $RUNDIR/scan"
+            if [ -d "$RUNDIR/bdio" ]
+            then
+                rm -rf "$RUNDIR/bdio"
+                msg "Deleting $RUNDIR/bdio"
+            fi
+            if [ -d "$RUNDIR/extractions" ]
+            then
+                rm -rf "$RUNDIR/extractions"
+                msg "Deleting $RUNDIR/extractions"
+            fi
+            if [ -d "$RUNDIR/scan" ]
+            then
+                rm -rf "$RUNDIR/scan"
+                msg "Deleting $RUNDIR/scan"
+            fi
         fi
     fi
 }
@@ -1055,12 +975,12 @@ run_report() {
             then
                 if [ $MODE_REPORT -eq 1 ]
                 then
-                    echo -n "	Component: '${COMPNAME/\'}' Policies Violated: "
+                    echo -n "	Component: '$COMPNAME' Policies Violated: "
                 fi
                 if [ $MODE_TESTXML -eq 1 ]
                 then
                     echo "<testcase name='${COMPNAME/\'}'>" >>$XMLPOL
-                    echo -n "<error message='${COMPNAME/\'} violates the following policies: " >>$XMLPOL
+                    echo -n "<error message='${COMPNAME/\'}' violates the following policies: " >>$XMLPOL
                 fi
                 api_call ${COMPURL}/policy-rules
                 if [ $? -ne 0 ]
@@ -1068,8 +988,8 @@ run_report() {
                     continue
                 fi
             
-                local POLNAMES=$($JQ -r '.items[].name' $TEMPFILE 2>/dev/null | tr '\n' '|' | sed -e "s/'/\"/g" )
-                local POLSEVERITIES=$($JQ -r '.items[].severity' $TEMPFILE 2>/dev/null | tr '\n' ',' | sed -e "s/'/\"/g" )
+                local POLNAMES=$($JQ -r '.items[].name' $TEMPFILE 2>/dev/null | tr '\n' '|' | escape_string )
+                local POLSEVERITIES=$($JQ -r '.items[].severity' $TEMPFILE 2>/dev/null | tr '\n' ',' | escape_string )
                 IFS='|'
                 sevind=1
                 for polname in $POLNAMES
@@ -1347,13 +1267,143 @@ then
     end 1
 fi
 
-process_args $*
+getargval() {
+    local ARG=$(echo "$*" | cut -f2 -d=)
+    ARG="${ARG#\'}"
+    ARG="${ARG%\'}"
+    if [ ! -z "$(echo $*|grep ' ')" ]
+    then
+        echo "'${ARG}'"
+    else
+        echo "${ARG}"
+    fi
+}
+
+procarg() {
+    local OPT=$(echo "$*" | cut -f1 -d=)
+    echo "${OPT}=$(getargval $*)"
+}
+
+# Process arguments
+while (( "$#" )); do
+#     echo "processing '$1'"
+    case "$1" in
+# Ignored arguments
+        --blackduck.offline.mode=*)
+            shift; continue
+            ;;
+        --detect.blackduck.signature.scanner.host.url=*)
+            shift; continue
+            ;;
+# Arguments NOT to be passed to detect.sh
+        --report)
+            debug "process_args(): MODE_REPORT set"
+            MODE_REPORT=1
+            shift; continue
+            ;;
+        --quiet)
+            debug "process_args(): MODE_QUIET set"
+            MODE_QUIET=1
+            shift; continue
+            ;;
+        --markdown)
+            debug "process_args(): MODE_MARKDOWN set"
+            MODE_MARKDOWN=1
+            shift; continue
+            ;;
+        --file)
+            debug "process_args(): MODE_PREVFILE set"
+            MODE_PREVFILE=1
+            shift; continue
+            ;;
+        --testxml)
+            debug "process_args(): MODE_TESTXML set"
+            MODE_TESTXML=1
+            shift; continue
+            ;;
+        --reset)
+            debug "process_args(): MODE_RESET set"
+            MODE_RESET=1
+            shift; continue
+            ;;
+        --detectscript=*)
+            DETECT_SCRIPT=$(getarg "$1")
+            debug "process_args(): DETECT_SCRIPT set to $DETECT_SCRIPT"
+            if [ ! -r "$DETECT_SCRIPT" ]
+            then
+                error "Detect script $DETECT_SCRIPT does not exist"
+            fi
+            shift; continue
+            ;;
+        --sigtime=*)
+            SIGTIME=$(getarg "$1")
+            debug "process_args(): SIGTIME set to $SIGTIME"
+            shift; continue
+            ;;
+        --curlopts=*)
+            CURLOPTS=$(getarg "$1")
+            shift; continue
+            ;;
+# Unsupported arguments
+        --detect.blackduck.signature.scanner.snippet.matching=*|--detect.blackduck.signature.scanner.upload.source.mode=*|--detect.blackduck.signature.scanner.copyright.search=*|--detect.blackduck.signature.scanner.license.search=*|--detect.binary.scan.*)
+            debug "process_args(): unsupported option"
+            UNSUPPORTED=1
+            shift; continue
+            ;;
+# Arguments to be passed to detect.sh
+        --blackduck.api.token=*)
+            debug "process_args(): BLACKDUCK_API_TOKEN identified from command line option"
+            API_TOKEN="$(getargval $1)=="
+            ;;
+        --blackduck.url=*)
+            debug "process_args(): BLACKDUCK_URL identified from command line option"
+            BD_URL=$(getargval "$1")
+            ;;
+        --detect.project.name=*)
+            DETECT_PROJECT=1
+            ;;
+        --detect.project.version.name=*)
+            DETECT_VERSION=1
+            ;;
+        --spring.profiles.active=*)
+            local YML="application-$(getargval $1).yml"
+            if [ ! -r $YML ]
+            then
+                YML=
+            fi
+            DETARGS="$DETARGS $(procarg $1)"
+            ;;
+        --blackduck.timeout=*|--detect.force.success=*|--detect.notices.report=*|--detect.policy.check.fail.on.severities=*|--detect.risk.report.pdf=*|--detect.wait.for.results=*)
+            DETECT_ACTION=1
+            msg "Detect Action identified - will rerun Detect after upload and scan completion"
+            debug "process_args(): Identified action argument $arg"
+            DETARGS="$DETARGS $(procarg $1)"
+            ;;
+        --detect.source.path=*)
+            SCANLOC=$(getarg $1)
+            SCANLOC=$(cd "$SCANLOC" 2>/dev/null; pwd)
+            ;;
+        --*)
+            ;;
+      esac
+      DETARGS="$DETARGS $(procarg $1)"
+      shift
+done
+
+# echo "DETARGS = '$DETARGS'"
+debug "Args processed"
+
+check_env()
+
+if [ $UNSUPPORTED -eq 1 ]
+then
+    error "Unsupported Detect options specified (Snippet or Binary)"
+fi
+
 if [ -z "$API_TOKEN" -o -z "$BD_URL" ]
 then
     error "No connection data for BD Server (BLACKDUCK_URL or BLACKDUCK_API_TOKEN)"
 fi
-
-debug "Args processed"
 
 TOKEN=$(get_token)
 
@@ -1421,7 +1471,7 @@ msg "Checking for signature scan ..."
 SIGDATE=$(check_sigscan $SIGTIME)
 if [ $? -eq 1 ]
 then
-    CLNAME=$(proc_sigscan | sed -e 's/ /%20/g' -e 's/"//g')
+    CLNAME=$(proc_sigscan | escape_string)
     if [ $? -ne 0 ]
     then
         msg "Unable to upload sig scan json"
@@ -1458,19 +1508,19 @@ then
         wait_for_scans "${BD_URL}/api/codelocations?q=name:${CLNAME}"
         if [ $? -ne 0 ]
         then
-            error "wait_for_scans() for sig scan returned error"
+            error2 "wait_for_scans() for sig scan returned error"
         fi
     fi
     debug "Waiting for version scans ..."
     wait_for_scans "${VERURL//\"}/codelocations"
     if [ $? -ne 0 ]
     then
-        error "wait_for_scans() for version returned error"
+        error2 "wait_for_scans() for version returned error"
     fi
     wait_for_bom_completion $VERURL
     if [ $? -ne 0 ]
     then
-        error "wait_for_bom_completion() returned error"
+        error2 "wait_for_bom_completion() returned error"
     fi
     if [ $DETECT_ACTION -eq 1 ]
     then
